// BulletTimeStudio/Cloud/admin/src/app/pages/common/socket.interface.ts
export enum SOCKET_EVENTS {
    EVENT_CAMERA_STATUS = 'camera_status',
    EVENT_CAMERA_STATUS_FROM_EDGE = 'camera_status_from_edge',
    START_RECORDING = 'start_recording',
    STOP_RECORDING = 'stop_recording',
    EVENT_CMD = 'cmd',
}

export enum SOCCER_EVENT {
    EVENT_HEARTBEAT = 'heartbeat',
    EVENT_DEVICEOFFLINE = 'device_offline',
}

export enum BASKETBALL_EVENT {
    REBOOT_DEVICE_SUCCESS_CLOUD = 'reboot_device_success_cloud',
    REBOOT_DEVICE_FAIL_CLOUD = 'reboot_device_fail_cloud',
    REBOOT_DEVICE_CLOUD = 'reboot_device_cloud',
}